webserver/app/requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-dateutil==2.7.1
pytz==2018.3
six==1.11.0
Werkzeug==0.14.1
